refactor: swagger spec interface 분리
